pyproject.toml
[project]
name = "labs"
version = "0.1.0"
description = "Mini console app"
requires-python = ">=3.14"
dependencies = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "structlog>=25.5.0",
    "typer>=0.20.0",
]
